requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
distro==1.8.0
Django==4.1.7
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
environ==1.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.6.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==3.2.0
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.0
paramiko==3.1.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
texttable==1.6.7
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==0.59.0
